feat: center the image and text vertically, bug squashing, style and minor cleanup
I experienced random crashes when 9weather downloaded the icon and
was trying to convert it to a plan9 image, turns out that the culprit
was the read(2) call to /mnt/web/n/body, I originally just used
readn(2) there and in hindsight it seems it was better suited. From
the man page: Readn is just like read, but does successive read
calls until nbytes have been read, or a read system call returns a
non-positive count.