feat: use environment variables and true color images
We are now able to set the API key with the `openweathermap`
environment variable, and we can also set the ZIP/city with the
`ZIP` variable.

According to the OpenWeather API the icon changes during night time
to greyscaled ones. While I was looking through the code I noticed
that I had shuffled the arguments for the `draw()` around which
made them greyscaled all the time.

Reference: https://openweathermap.org/weather-conditions#Icon-list